chore(souffle): `executor` module
